+++ tests/unit/Codeception/Module/WPFilesystemTest.php
@@ -826,6 +826,38 @@ PHP;
         Assert::assertFileDoesNotExist($pluginFolder);
     }
 
+    /**
+     * It should allow having a single file plugin with code
+     * @test
+     */
+    public function it_should_allow_having_a_single_file_plugin_with_code()
+    {
+        $sut = $this->make_instance();
+
+        $pluginsFolder = $this->config['wpRootFolder'] . $this->config['plugins'];
+        $pluginFile =  $pluginsFolder . '/plugin.php';
+
+        $code = "echo 'Hello world';";
+        $sut->havePlugin('plugin.php', $code);
+
+        $this->assertFileExists($pluginFile);
+
+        $expected = <<<PHP
+<?php
+/*
+Plugin Name: plugin
+Description: plugin
+*/
+
+echo 'Hello world';
+PHP;
+        $this->assertEquals(normalizeNewLine($expected), normalizeNewLine(file_get_contents($pluginFile)));
+
+        $sut->_after($this->stubProphecy(TestInterface::class)->reveal());
+
+        Assert::assertFileDoesNotExist($pluginFile);
+        $this->assertFileExists($pluginsFolder);
+    }
 
     /**
      * It should allow having a mu-plugin with code
@@ -1061,5 +1093,48 @@ PHP;
         Assert::assertFileDoesNotExist($themeStyleFile);
         Assert::assertFileDoesNotExist($themeIndexFile);
         Assert::assertFileDoesNotExist($themeFunctionsFile);
+    }
+
+    public function plugin_path_with_diff_path_separators_data_provider()
+    {
+        return [
+            '/' => ['foo/plugin.php'],
+            '\\' => ['foo\\plugin.php']
+        ];
+    }
+
+    /**
+     * It should allow using different directory separators to havePlugin
+     *
+     * @test
+     * @dataProvider plugin_path_with_diff_path_separators_data_provider
+     */
+    public function should_allow_using_different_directory_separators_to_have_plugin($pluginPath)
+    {
+        $sut = $this->make_instance();
+
+        $pluginsFolder = $this->config['wpRootFolder'] . $this->config['plugins'];
+        $pluginFile =  $pluginsFolder . '/foo/plugin.php';
+
+        $code = "echo 'Hello world';";
+        $sut->havePlugin($pluginPath, $code);
+
+        $this->assertFileExists($pluginFile);
+
+        $expected = <<<PHP
+<?php
+/*
+Plugin Name: foo
+Description: foo
+*/
+
+echo 'Hello world';
+PHP;
+        $this->assertEquals(normalizeNewLine($expected), normalizeNewLine(file_get_contents($pluginFile)));
+
+        $sut->_after($this->stubProphecy(TestInterface::class)->reveal());
+
+        Assert::assertFileDoesNotExist($pluginFile);
+        $this->assertFileExists($pluginsFolder);
     }
 }